Added a Debug Mode
The debug mode allows us to generate the execution files (batch or regular sh) without submitting the files for execution. As the name suggests, this mode is designed to help us debug the test generation process. It also allows us to check aspects that do not depend on the execution of the tests.