Fixing local execution & rolling back to default IOR output
In local execution, we can only vary the file size and the number of processes. Since we are currently considering only the computing nodes, striping, and file size tests, we need to limit the tests for local execution to only the file size.

Moreover, I realize that the CSV generated by IOR via the -O option does not contain all the information we need. Therefore, I am changing it to generate the default file and using the ior_2_csv.py script to capture all the information we need when generating the CSV