Adding build method to BenchmarkRunner
Added a build method to automate class selection based on the benchmark defined in the config file.

Removed file and folder creation from the IORBenchmark class. This logic is now handled by the planner, since the infrastructure (including folder and file definitions) should be managed by the core of IOPS, not be benchmark-dependent.

Updated the parse_output function to return all metrics from the summary file.